Completed communication

+++ partner/src/main/java/hu/otp/mobile/partner/service/EventService.java
@@ -7,7 +7,18 @@ import otp.mobile.backend.common.domain.EventSeating;
 
 public interface EventService {
 
-	public EventSeating getEvent(int eventId);
+	/**
+	 * Queries the event details
+	 * 
+	 * @param eventId the event ID, not <code>null</code>
+	 * @return returns the event's seating details
+	 */
+	public EventSeating getEvent(Long eventId);
 
+	/**
+	 * Queries the list of events
+	 * 
+	 * @return an array of less verbose event data
+	 */
 	public List<Event> getEvents();
 }
