+++ partner/src/main/java/hu/otp/mobile/partner/service/ReservationService.java
@@ -4,5 +4,12 @@ import opt.mobile.backend.common.dto.ReservationResult;
 
 public interface ReservationService {
 
-	public ReservationResult reserve(int eventId, int seatId);
+	/**
+	 * Attempts to reserve a seat for the specific event
+	 * 
+	 * @param eventId the event ID, not <code>null</code>
+	 * @param seatId the seat's ID, not <code>null</code>
+	 * @return the reservation result dto
+	 */
+	public ReservationResult reserve(Long eventId, Long seatId);
 }
